refactor: Option, Options 도메인 패키지 통합
