--- backend/src/main/resources/static/mapper/options/options_mapper.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.h2o.h2oServer.domain.options.mapper.OptionsMapper">
-    <select id="findTrimPackages" resultType="TrimExtraOptionEntity">
-        select id, name, image, pc_detail_image as pc_image, mobile_detail_image as mobile_image, category, choice_ratio, price
-        from trims_package
-                join package
-                on trims_package.package_id = package.id
-        where trim_id = #{trimId}
-        order by package_id;
-    </select>
-    
-    <select id="findTrimPackagesWithRange" resultType="TrimExtraOptionEntity">
-        select id, name, image, pc_detail_image as pc_image, mobile_detail_image as mobile_image, category, choice_ratio, price
-        from trims_package
-            join package
-                on trims_package.package_id = package.id and
-                   trims_package.trim_id = #{trimId}
-        order by package_id
-        <if test="pageRange.from != null and pageRange.count != null">
-            limit #{pageRange.count}
-            offset #{pageRange.from}
-        </if>
-    </select>
-
-    <select id="findTrimExtraOptions" resultType="TrimExtraOptionEntity">
-        select id, name, image, pc_detail_image as pc_image, mobile_detail_image as mobile_image, category, choice_ratio, price
-        from trims_options
-                join options
-                on trims_options.option_id = options.id
-        where trim_id = #{trimId} and option_type = 'extra'
-        order by option_id;
-    </select>
-
-    <select id="findTrimExtraOptionsWithRange" resultType="TrimExtraOptionEntity">
-        select id, name, image, pc_detail_image as pc_image, mobile_detail_image as mobile_image, category, choice_ratio, price
-        from trims_options
-            join options
-                on trims_options.option_id = options.id and
-                   trims_options.trim_id = #{trimId} and
-                   trims_options.option_type = 'extra'
-        order by option_id
-        <if test="pageRange.from != null and pageRange.count != null">
-            limit #{pageRange.count}
-            offset #{pageRange.from}
-        </if>
-    </select>
-
-    <select id="findTrimDefaultOptions" resultType="TrimDefaultOptionEntity">
-        select id, name, image, pc_detail_image as pc_image, mobile_detail_image as mobile_image, category, choice_ratio, use_count
-        from trims_options
-            join options
-                on trims_options.option_id = options.id
-        where trim_id = #{trimId} and option_type = 'default'
-        order by option_id;
-    </select>
-
-    <select id="findTrimDefaultOptionsWithRange" resultType="TrimDefaultOptionEntity">
-        select id, name, image, pc_detail_image as pc_image, mobile_detail_image as mobile_image, category, choice_ratio, use_count
-        from trims_options
-            join options
-                on trims_options.option_id = options.id and
-                   trims_options.trim_id = #{trimId} and
-                   trims_options.option_type = 'default'
-        order by option_id
-        <if test="pageRange.from != null and pageRange.count != null">
-            limit #{pageRange.count}
-            offset #{pageRange.from}
-        </if>
-    </select>
-
-    <select id="findPackageHashTags" resultType="HashTagEntity">
-        select name
-        from package_hashtag
-                join hashtag
-                on package_hashtag.hashtag_id = hashtag.id
-        where package_id = #{packageId}
-        order by hashtag_id
-        limit 5;
-    </select>
-
-    <select id="findOptionHashTag" resultType="HashTagEntity">
-        select name
-        from options_hashtag
-                join hashtag
-                on options_hashtag.hashtag_id = hashtag.id
-        where option_id = #{optionId}
-        order by hashtag_id
-        limit 5;
-    </select>
-
-    <select id="findTrimPackageSize" resultType="long">
-        select count(*)
-        from trims_package
-            join package
-                on trims_package.package_id = package.id and
-                   trim_id = #{trimId};
-    </select>
-
-    <select id="findTrimOptionSize" resultType="long">
-        select count(*)
-        from trims_options
-                 join options
-                      on trims_options.option_id = options.id and
-                         trims_options.trim_id = #{trimId} and
-                         trims_options.option_type = #{optionType.label};
-    </select>
-</mapper>
